% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_clock_drift.R
\name{clockDriftUI}
\alias{clockDriftUI}
\title{Clock Drift Correction Module}
\usage{
clockDriftUI(id)
}
\arguments{
\item{id}{Module ID}

\item{heat_pulse_data}{Reactive containing heat_pulse_data object}
}
\value{
Reactive containing corrected data (or NULL if no correction applied)
}
\description{
Shiny module for correcting clock drift in heat pulse data
Assumes first pulse time is correct (synced at start)
}
