% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_data_upload.R
\name{dataUploadUI}
\alias{dataUploadUI}
\title{Data Upload Module}
\usage{
dataUploadUI(id)
}
\arguments{
\item{id}{Module ID}
}
\value{
Reactive containing heat_pulse_data object
}
\description{
Shiny module for uploading heat pulse data files
}
