% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_number}
\alias{format_number}
\title{Format large numbers with thousands separator}
\usage{
format_number(x, digits = 0)
}
\arguments{
\item{x}{Numeric value}

\item{digits}{Number of decimal places}
}
\value{
Formatted character string
}
\description{
Format large numbers with thousands separator
}
\keyword{internal}
