% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_vh}
\alias{format_vh}
\title{Format Vh value for display}
\usage{
format_vh(vh, digits = 2)
}
\arguments{
\item{vh}{Velocity value in cm/hr}

\item{digits}{Number of decimal places}
}
\value{
Formatted string with units
}
\description{
Format Vh value for display
}
\keyword{internal}
