% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_date_range_text}
\alias{get_date_range_text}
\title{Utility Functions for Shiny Sap Flow App}
\usage{
get_date_range_text(measurements)
}
\arguments{
\item{measurements}{Tibble with datetime column}
}
\value{
Character string with formatted date range
}
\description{
Helper functions used across modules
}
\details{
Get formatted date range text from measurements
}
\keyword{internal}
