% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_duration_text}
\alias{get_duration_text}
\title{Get formatted duration text from measurements}
\usage{
get_duration_text(measurements)
}
\arguments{
\item{measurements}{Tibble with datetime column}
}
\value{
Character string with formatted duration
}
\description{
Get formatted duration text from measurements
}
\keyword{internal}
