% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_method_colours}
\alias{get_method_colours}
\title{Create method colour palette}
\usage{
get_method_colours(methods)
}
\arguments{
\item{methods}{Character vector of method names}
}
\value{
Named vector of colours
}
\description{
Create method colour palette
}
\keyword{internal}
