% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_methods_from_results}
\alias{get_methods_from_results}
\title{Extract method names from vh_results}
\usage{
get_methods_from_results(vh_results)
}
\arguments{
\item{vh_results}{Results tibble from calc_heat_pulse_velocity}
}
\value{
Character vector of unique method names
}
\description{
Extract method names from vh_results
}
\keyword{internal}
