% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_quality_colour}
\alias{get_quality_colour}
\title{Get quality flag colour}
\usage{
get_quality_colour(flag)
}
\arguments{
\item{flag}{Quality flag ("OK", "WARNING", "ERROR", "SUSPECT")}
}
\value{
Colour code (hex)
}
\description{
Get quality flag colour
}
\keyword{internal}
