% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_quality_icon}
\alias{get_quality_icon}
\title{Get quality flag icon}
\usage{
get_quality_icon(flag)
}
\arguments{
\item{flag}{Quality flag}
}
\value{
Icon name for shiny::icon()
}
\description{
Get quality flag icon
}
\keyword{internal}
