% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{help_section}
\alias{help_section}
\title{Create a collapsible help section}
\usage{
help_section(title, content)
}
\arguments{
\item{title}{Section title}

\item{content}{HTML content}
}
\value{
Shiny UI element
}
\description{
Create a collapsible help section
}
\keyword{internal}
