% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_valid_datetime}
\alias{is_valid_datetime}
\title{Validate datetime input}
\usage{
is_valid_datetime(datetime)
}
\arguments{
\item{datetime}{POSIXct datetime}
}
\value{
Logical - TRUE if valid
}
\description{
Validate datetime input
}
\keyword{internal}
