% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_methods.R
\name{methodsUI}
\alias{methodsUI}
\title{Method Selection and Calculation Module}
\usage{
methodsUI(id)
}
\arguments{
\item{id}{Module ID}

\item{heat_pulse_data}{Reactive containing corrected heat_pulse_data}

\item{probe_config}{Reactive containing probe configuration}

\item{wood_properties}{Reactive containing wood properties}
}
\value{
Reactive containing vh_results
}
\description{
Shiny module for selecting HPV calculation methods and running calculations
}
