% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notify_helper.R
\name{notify_progress}
\alias{notify_progress}
\title{Show progress notification (no buttons, no auto-close)}
\usage{
notify_progress(title, text)
}
\description{
Show progress notification (no buttons, no auto-close)
}
