% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notify_helper.R
\name{notify_warning}
\alias{notify_warning}
\title{Show warning notification with auto-close}
\usage{
notify_warning(session, title, text, timer = 3000)
}
\description{
Show warning notification with auto-close
}
