% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{package_available}
\alias{package_available}
\title{Check if package is available}
\usage{
package_available(pkg)
}
\arguments{
\item{pkg}{Package name}
}
\value{
Logical
}
\description{
Check if package is available
}
\keyword{internal}
