% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_probe_vertical}
\alias{plot_probe_vertical}
\title{Plot Probe Configuration - Vertical View}
\usage{
plot_probe_vertical(validation)
}
\arguments{
\item{validation}{List from validate_probe_tree_config()}
}
\value{
ggplot2 object
}
\description{
Creates a side/vertical view of probes inserted into stem
}
\keyword{internal}
