% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_plot_timeseries.R
\name{plotTimeseriesUI}
\alias{plotTimeseriesUI}
\title{Time Series Visualization Module}
\usage{
plotTimeseriesUI(id)
}
\arguments{
\item{id}{Module ID}

\item{vh_results}{Reactive containing vh_results tibble}
}
\value{
None (displays plot)
}
\description{
Shiny module for interactive plotly time series of HPV results
}
