% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_pulse_trace.R
\name{pulseTraceUI}
\alias{pulseTraceUI}
\title{Pulse Trace Viewer Module}
\usage{
pulseTraceUI(id)
}
\arguments{
\item{id}{Module ID}

\item{heat_pulse_data}{Reactive containing heat_pulse_data}

\item{selected_pulse_id}{Reactive containing selected pulse ID from plot click}
}
\value{
None (displays pulse trace plot)
}
\description{
Shiny module for viewing individual pulse temperature traces
}
